Add files via upload
Added two more Batch scripts for helping with Video stuff.

Both require ffmpeg to be on the system in the path.

4K 2 1080p is one I wrote to go through a directory and all the ones below it, and make good enough reduced quality versions. It still requires you to manually replace the originals. So it shouldn't erase anything on you by accident. I use this on all my Phone footage. I believe that the 4k content from phones is much better. Both in resolution and bitrate. But I don't think 4k content makes much sense. So I am converting all that phone video to 1080p. It looks very sharp, and ends up being about half the size and is good enough for what I need phone content for. (This is because I back up my phone photos to the cloud)

It's not a very complicated script, and ideally it would be able to take a file list, target resolution information, and compression info to allow better flexibility. But since I always target the same quality i didn't do that.

count_4k_vids is the second script. It's more or less the same code. It's name is also a misnomer. It really is just any content that is bigger then 1080p on the short edge. It runs much faster, so I use it to just see how many files I need to convert. Also lets me see if maybe some of my assumptions for the script are wrong without having to let the whole thing run.